07_rag/requirements.txt
langgraph
langsmith
langchain_anthropic
tavily-python
langchain_community
langchain-text-splitters 
bs4